requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
